Renamed DomainCertificate to CertificateValueObjects
