+++ domains/certificates/Certificates.sln
@@ -27,9 +27,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Worker.UnitTests", "Registr
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "API.IntegrationTests", "Query.API\API.IntegrationTests\API.IntegrationTests.csproj", "{741B4018-7BB1-46EF-A285-87EC11418FC5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DomainCertificate", "Shared\DomainCertificate\DomainCertificate.csproj", "{1F343067-3640-4209-9D7B-33D2B2E26033}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DomainCertificate.Tests", "DomainCertificate.Tests\DomainCertificate.Tests.csproj", "{59127B55-9983-4BEB-9FC6-283F485B9C26}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CertificateValueObjects", "Shared\CertificateValueObjects\CertificateValueObjects.csproj", "{FFBE8D8C-4504-4017-847C-F21C5273673E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -73,14 +71,10 @@ Global
 		{741B4018-7BB1-46EF-A285-87EC11418FC5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{741B4018-7BB1-46EF-A285-87EC11418FC5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{741B4018-7BB1-46EF-A285-87EC11418FC5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1F343067-3640-4209-9D7B-33D2B2E26033}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1F343067-3640-4209-9D7B-33D2B2E26033}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1F343067-3640-4209-9D7B-33D2B2E26033}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1F343067-3640-4209-9D7B-33D2B2E26033}.Release|Any CPU.Build.0 = Release|Any CPU
-		{59127B55-9983-4BEB-9FC6-283F485B9C26}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{59127B55-9983-4BEB-9FC6-283F485B9C26}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{59127B55-9983-4BEB-9FC6-283F485B9C26}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{59127B55-9983-4BEB-9FC6-283F485B9C26}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FFBE8D8C-4504-4017-847C-F21C5273673E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FFBE8D8C-4504-4017-847C-F21C5273673E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FFBE8D8C-4504-4017-847C-F21C5273673E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FFBE8D8C-4504-4017-847C-F21C5273673E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -95,8 +89,7 @@ Global
 		{005E4EDB-9F14-4CDC-A2A9-C5D261872E5C} = {9AB85139-82C5-45D0-97CD-CD9FCADB6380}
 		{3FC7672A-5A92-4B96-94E3-60B7468148FD} = {9AB85139-82C5-45D0-97CD-CD9FCADB6380}
 		{741B4018-7BB1-46EF-A285-87EC11418FC5} = {C500AE25-2567-4855-863B-7B51BC59F806}
-		{1F343067-3640-4209-9D7B-33D2B2E26033} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
-		{59127B55-9983-4BEB-9FC6-283F485B9C26} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
+		{FFBE8D8C-4504-4017-847C-F21C5273673E} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8710F924-9EE6-4142-A232-456121E239B0}
