+++ domains/certificates/Query.API/API.IntegrationTests/CertificatesByOwnerProjectionTest.cs
@@ -5,8 +5,7 @@ using API.IntegrationTests.Testcontainers;
 using API.Query.API.Projections;
 using API.Query.API.Projections.Views;
 using CertificateEvents.Aggregates;
-using DomainCertificate.Primitives;
-using DomainCertificate.ValueObjects;
+using CertificateValueObjects;
 using FluentAssertions;
 using Marten;
 using VerifyXunit;
