+++ domains/certificates/Query.API/API.UnitTests/DataSyncSyncer/Client/DataSyncClientTest.cs
@@ -6,7 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using API.DataSyncSyncer.Client;
 using API.DataSyncSyncer.Client.Dto;
-using DomainCertificate.ValueObjects;
+using CertificateValueObjects;
 using FluentAssertions;
 using MeasurementEvents;
 using Microsoft.Extensions.Logging;
