+++ domains/certificates/Query.API/API/API.csproj
@@ -32,8 +32,8 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Shared\AggregateRepositories\AggregateRepositories.csproj" />
     <ProjectReference Include="..\..\Shared\CertificateEvents\CertificateEvents.csproj" />
+    <ProjectReference Include="..\..\Shared\CertificateValueObjects\CertificateValueObjects.csproj" />
     <ProjectReference Include="..\..\Shared\Contracts\Contracts.csproj" />
-    <ProjectReference Include="..\..\Shared\DomainCertificate\DomainCertificate.csproj" />
     <ProjectReference Include="..\..\Shared\MeasurementEvents\MeasurementEvents.csproj" />
   </ItemGroup>
