+++ domains/certificates/Query.API/API/ContractService/ContractServiceImpl.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using API.ContractService.Clients;
 using API.ContractService.Repositories;
-using DomainCertificate.Primitives;
+using CertificateValueObjects;
 using Marten.Exceptions;
 using static API.ContractService.CreateContractResult;
