+++ domains/certificates/Query.API/API/GranularCertificateIssuer/EnergyMeasuredEventHandler.cs
@@ -3,10 +3,8 @@ using System.Threading.Tasks;
 using AggregateRepositories;
 using API.ContractService;
 using CertificateEvents.Aggregates;
+using CertificateValueObjects;
 using Contracts.Certificates;
-using DomainCertificate;
-using DomainCertificate.Primitives;
-using DomainCertificate.ValueObjects;
 using MassTransit;
 using MeasurementEvents;
 using Microsoft.Extensions.Logging;
