+++ domains/certificates/Shared/CertificateValueObjects/CertificateValueObjects.csproj
