+++ domains/certificates/Shared/CertificateValueObjects/Technology.cs
@@ -1,4 +1,4 @@
-namespace DomainCertificate.Primitives;
+namespace CertificateValueObjects;
 
 public record Technology(
     string FuelCode, // stamdata
