+++ domains/certificates/Shared/CertificateValueObjects/ValueObject.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace DomainCertificate;
+namespace CertificateValueObjects;
 
 /// <summary>
 /// Taken from Microsoft: https://learn.microsoft.com/en-us/dotnet/architecture/microservices/microservice-ddd-cqrs-patterns/implement-value-objects
