+++ domains/certificates/Shared/Contracts/Contracts.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\DomainCertificate\DomainCertificate.csproj" />
+    <ProjectReference Include="..\CertificateValueObjects\CertificateValueObjects.csproj" />
   </ItemGroup>
 
 </Project>
